[POA-2713] Implement Run function (#78)
In this PR, we have implemented the main run function which will start
all the workers.

* Start all the required workers
* Call the `StartProcessInExistingPods` function to check for existing
pods
* Start listening to the interrupt signal to stop the process
* On interrupt signal workers to stop
* Stop running apidump processes
* Added 2 new functions
  * Dump the pod states on regular intervals if debug mode enabled
  * prune stopped processes from map
* Also moved some functions to their separate files
